Update README.md
The list that I put are my favorite TV series. The first series is the Supernaturals it's about a two brother fighting or hunting devils or monster and solving supernatural cases. The second is The Boys which is about a group of regular people fighting super heroes. The third is Lucifer it's about the demon lucifer who came to earth and try to solve cases with the female officer.